+++ src/instant_pattern.cpp
@@ -6,8 +6,8 @@ unsigned long InstantPattern::previousMillis = 0;
 
 void InstantPattern::run()
 {
-    int pattern = PatternManager::getCurrentPattern();
-    int step = PatternManager::getCurrentStep();
+    uint8_t pattern = PatternManager::getCurrentPattern();
+    uint8_t step = PatternManager::getCurrentStep();
     const Step currentPatternStep = patterns[pattern][step];
 
     if (millis() - previousMillis < currentPatternStep.duration)
@@ -18,7 +18,7 @@ void InstantPattern::run()
 
     previousMillis = millis();
 
-    for (int i = 0; i < LED_COUNT; i++)
+    for (uint8_t i = 0; i < LED_COUNT; i++)
     {
         uint8_t brightness = currentPatternStep.leds[i];
         DUTY_TYPE gammaValue = PGM_READ_DUTY(&gammaMap[brightness]);
